Add automatic reconnection and unify image source
Note that the automatic reconnection might not work as expected if there
are muptiple cameras connected to the system. OpenCV addresses cameras
by their index and certain backends change the indices of already
connected cameras when a another camera is connected or disconnected.
One notable example of such a backend is AVFoundation on macOS.