Add 'smooth' camera config option
When the 'smooth' option is set to a value in > 0.0, the retrieved
camera image will be combined with the previous frames using a weighted
average: image = prev_image * smooth + cur_image * (1.0 - smooth)

This helps reducing noise at the expense of a slightly delayed
detection. But since the camera most likely grabs frames at a much
higher rate than the actual detection rate, values of 0.5 to 0.75 still
seem to be usable in practice.