Add 'f' hotkey to freeze input image
The camera video stream usually contains a lot of noise. When
calibrating on a remote host, this leads to a lot of data being sent
over the network, even though a static image would be good enough for
calibration. Freezing the image reduces the encoding load on the remote
host and lowers the required network bandwidth.