Replace `findHomography` by `getPerspectiveTransform`
A homography matrix between two 2D spaces is defined by two matching
sets of excactly four points. This is the configuration we aim to
reproduce using the physical markers, so `getPerspectiveTransform` is
sufficient and `findHomography` is actually overkill since it is meant
to be used for larger sets of sometimes inaccurate points.