Add marker redetection heuristic to improve performance
Marker positions usually do not change much. Therefore, searching the
whole image in each iteration is a waste of ressources most of the time.
This patch adds a heuristic that first tries to find each marker in the
vicinity of its previous location. If not all markers are found, the
whole image is searched. The search area can be enlarged by a
user-provided tolerance parameter.