Add `minConstrast` option to aid empty tray detection
Empty trays can still be classified as valid tags when adaptive
thresholding is applied to them. This is because such images still don't
have a uniform color/gray value, but rather have values in a narrow
range due to noise or lighting conditions. This patch introduces a
`minConstrast` option that makes the detector reliably classify all
tiles having a contrast below `minContrast` as `unknown`.