Fix computation of centers of projectively distorted markers
The center of a square marker is just the arithmetic mean of its
corner's coordinates. However, the image of a marker as seen through a
camera is projectively distorted such that this computation gives wrong
results most of the time. The center in the image is actually given by
the intersection of the diagonals of the marker's quadrilateral, which
is what this patch implements.