Fix mapping of slice-local marker coordinates to global coordinates
To speed up the marker tracking, the marker detection is first executed
on small slices of the image centered at the previous marker positions.
Since these slices are slightly larger than the previously detected
markers, their bounding coordinates need to be clipped to the image
boundaries. This needs to be taken into account, when transforming
marker coordinates from its slice to the whole image.