Adjust ROI vertices relative to the current view
Prior to this patch, the WASD keys would move the ROI relative to the
originally configured coordinate system. In a marker-based setup, this
can be very tedious and confusing when the current view is a heavily
transformed one e.g. due to camera rotation. With this patch, the
the current marker positions and ROI vertices are made the new source
coordinate system such that ROI vertex adjustments happen relative to
the current view. Repeating this transformation process many times can
lead to drift (due to numerical errors) in the ROI vertices, which are
not currently modified by user input. But since this happens during
calibration by the user, the drift is directly observable and can be
corrected by the user.