Fix marker config export when marker centers are not defined
Specifying a marker config that just contains the marker dictionary id
and the marker ids is valid even if the marker centers are not included.
This is especially useful for initial calibration. This patch fixes a
missing type conversion when exporting the marker centers to a config
that does not contain them yet.