Adjust ROI in large steps using keys `ijkl` in addition to `WASD`
The processing of key events in OpenCV's HighGUI is very basic and does
not handle modifier keys particularly well. In addition, the behavior
is platform and UI toolkit dependend. Using the `WASD` keys (uppercase
`wasd`) for making big steps in the ROI adjustment never really worked
on Linux/QT and there doesn't seem to be a cross-platform/-toolkit way
of dealing with this in OpenCV. Therefore, this patch adds the keys
`ijkl` that serve the same purpose as `WASD`.