Refactor `CropTileCells` towards using `remap`
Previously, the margins of tile cells were removed by copying the inner
part of all cells into a new image. This was done with sub-pixel
accuracy since, sometimes, the amount of usable pixels is very low and
rounding to pixel boundaries during several steps in the pipeline can
lead to insufficient data for tag detection. This code turned out to be
much slower than precomputing a sampling map for the whole input image
that is then just applied using `remap`.

Speedup of `CropTileCells`: ~30-35x