Add free scaling parameter `alpha` to `calibration` configuration
By default, only the rectangle of valid pixels is contained in the
undistorted image. The new  `alpha` parameter now allows to control the
amount of invalid pixels. At `alpha=0.0`, only valid pixels are shown.
At `alpha=1.0`, the mappings of all pixels of the original image are
maintained. Values in between interpolate accordingly.